API

This page is a dump of all the docstrings found in the code. This will be cleaned up in the future and organized to be more accessible.

EKF.Measurement — Type

Measurement{Nₘ, T}

Abstract type describing what is being measured.

EKF.Observation — Type

Observation(meas::Measurement, cov::SMatrix)

A observation consists of the acutal measurement and its associated covariance. This is particularly useful for when you have multiple independent measurments in a EKF.

Example

Consider measuring the position of a robot: ... struct Position{T} <: EKF.Measurement{3, T} px::T; py::T; pz::T end ... We express a particular obervation of this measurement as: ... positionObservation = Observation(Position(0.,0.,0.), SA[1. 0 0; 0 1 0; 0 0 1;]) ...

EKF.getCovariance — Method

getCovariance(obs::Observation)

Useful helper function.

EKF.getMeasurement — Method

getMeasurement(obs::Observation)

Useful helper function.