EKF.jl Documentation

Index

EKF.ExtendedKalmanFilter

EKF.estimateState

EKF.simulate

Extended Kalman Filtering

EKF.ExtendedKalmanFilter — Type

struct ExtendedKalmanFilter{T}

Extended Kalman Filter struct. Stores the dynamics and measurement functions as well as their cooresponding covariances.

Arguments

Q::AbstractArray{T}: Dynamics noise covariance matrix, must be symmetric

R::AbstractArray{T}: Measurement noise covariance matrix, must be symmetric

process::Function: dynamics function, steps the system forward

measure::Function: measurement function

The process and measure function have the following forms:

function process(state::AbstractVector, input::AbstractVector)
    ...
    return new_state
end

function measure(state::AbstractVector)
    ...
    return [θ₁, θ₂]
end

EKF.estimateState — Function

estimateState(est_state::AbstractArray{T}, input::AbstractArray{T}, measurement::AbstractArray{T}, errorCov::AbstractArray{T}, ekf::ExtendedKalmanFilter{T}) where T

Estimate the state of the system specified by ekf. Returns the new state at time step $i+1$

Arguments

est_state::AbstractArray{T}: Estimated state at time step $i$

input::AbstractArray{T}: Control input at time step $i$

measurement::AbstractArray{T}: Measurment of state at time step $i$

errorCov::AbstractArray{T}: Covariance error

ekf::ExtendedKalmanFilter{T}: ExtendedKalmanFilter struct specifying the dynamics and process aswell as their covariances

The process and measure function have the following forms:

function process(state::AbstractVector, input::AbstractVector)
    ...
    return new_state
end

function measure(state::AbstractVector)
    ...
    return [θ₁, θ₂]
end

EKF.simulate — Function

simulate(initState::AbstractArray, initEstimate::AbstractArray, errorCov::AbstractArray, inputs::AbstractArray, numSteps::Int64, ekf::ExtendedKalmanFilter)

Simulates the system specified by ekf::ExtendedKalmanFilter over time horizon.

Arguments

initState::AbstractArray: Inital state at time step 0

initEstimate::AbstractArray: Inital "guess" of the state at time step 0

errorCov::AbstractArray: Error covariances at each time step

inputs::AbstractArray: List of inputs for each time step

numSteps::Int64: Covariance error

ekf::ExtendedKalmanFilter: ExtendedKalmanFilter struct specifying the dynamics and process as well as their covariances